[serve] Created a new file for defining DeploymentID and changed imports for some internal files
