+++ python/ray/serve/_private/deploy_utils.py
@@ -6,10 +6,10 @@ from typing import Any, Dict, Optional, Union
 
 import ray
 import ray.util.serialization_addons
-from ray.serve._private.common import DeploymentID
 from ray.serve._private.config import DeploymentConfig, ReplicaConfig
 from ray.serve._private.constants import SERVE_LOGGER_NAME
 from ray.serve._private.deployment_info import DeploymentInfo
+from ray.serve.common import DeploymentID
 from ray.serve.schema import ServeApplicationSchema
 
 logger = logging.getLogger(SERVE_LOGGER_NAME)
