+++ python/ray/serve/_private/local_testing_mode.py
@@ -9,7 +9,7 @@ from typing import Any, Callable, Coroutine, Dict, Optional, Tuple, Union
 
 import ray
 from ray import cloudpickle
-from ray.serve._private.common import DeploymentID, RequestMetadata
+from ray.serve._private.common import RequestMetadata
 from ray.serve._private.constants import (
     RAY_SERVE_RUN_SYNC_IN_THREADPOOL,
     SERVE_LOGGER_NAME,
@@ -18,6 +18,7 @@ from ray.serve._private.replica import UserCallableWrapper
 from ray.serve._private.replica_result import ReplicaResult
 from ray.serve._private.router import Router
 from ray.serve._private.utils import GENERATOR_COMPOSITION_NOT_SUPPORTED_ERROR
+from ray.serve.common import DeploymentID
 from ray.serve.deployment import Deployment
 from ray.serve.exceptions import RequestCancelledError
 from ray.serve.handle import (
