+++ python/ray/serve/_private/proxy_router.py
@@ -6,12 +6,13 @@ from starlette.requests import Request
 from starlette.routing import Route
 from starlette.types import Scope
 
-from ray.serve._private.common import ApplicationName, DeploymentID, EndpointInfo
+from ray.serve._private.common import ApplicationName, EndpointInfo
 from ray.serve._private.constants import SERVE_LOGGER_NAME
 from ray.serve._private.thirdparty.get_asgi_route_name import (
     RoutePattern,
     get_asgi_route_name,
 )
+from ray.serve.common import DeploymentID
 from ray.serve.handle import DeploymentHandle
 
 logger = logging.getLogger(SERVE_LOGGER_NAME)
